optimize: 1. check global lock when use @GlobalLock Independently. 2. Check global lock retry when conflict.
